Restructure XcodeProject; drop SwiftPM, fix #8
Instead of relying on SwiftPM to generate the xcodeproj and create targets for the libraries, we manually created an xcodeproj and added each library as a macOS framework target. This allows for the most flexibility.
Also, use Accio for dependency management. This allows for better scalability than SwiftPM as it does not destory the xcodeproj each time a new library or dependency is added.