Add BasicLinearPingPongTracker
Instead of creating fully abstract and generic Continuous- and RepeatingPingPongTrackers, we first create a simple ping pong tracker which perfectly matches the requirements of MrFlap, specifically the bars.
In the future, more abstract PingPongTrackers can be created as required.
This means, the other files inside the PingPong folder are currently unfinished & unused.