Replace separate strategies (path through bar, way to point) with a single, unified strategy
This allows the whole sequence calculation to be made in a single place, which allows for additional solutions that cannot be found by the split-strategy approach.
Also, the split-strategy approach did not work most of the time, because the distances from the player to the starting point of the path-through-bar-sequence were often too small.