Fully optimize tap prediction performance
In particular, remove all `array1 + array2` calls in performance-critical methods.
For example, replace `for elem in array + [value]` with an ugly `for i in 0 ... array.count` loop.
This relieves Swift from creating a temporary array and performing alloc/dealloc burden.

Now, the Tap Prediction performance is actually optimal – nearly all of the time is spent in `arc4random`.