Associate TapDebugInfo with taps
This allows to store debug information (in our case, Jumps) in RelativeTaps, ScheduledTaps and PerformedTaps for later debug-usage.
This is required because these Jumps cannot be reconstructed at a later point because the underlying physics values (slightly) change each frame. Therefore, the Jumps along with their physics values are directly stored inside the relevant taps.