Use AbsoluteTapSequence instead of RelativeTapSequence for tap scheduling
This is necessary because the reference point of the solution calculator could be different from the reference point of the tap scheduler. For example, when tap scheduling takes 10ms, scheduled taps are all shifted by 10ms because they were relative to an undefined reference point. This MR makes this reference point explicit.